futurembryo/pyproject.toml
[project]
name = "futurembryo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.86.0",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "mcp>=1.0.0",
]
